data/libexec: Add sudo wrapper and host runner
'host-runner' is a shell script meant to be symlinked against to execute
commands on the host from inside of an OCI container. It makes use of
'flatpak-spawn' with it's option --host that takes care of the plumbing
from a container to the host. The command itself can not be run as root
because it relies on session bus, so instead the command, that is to be
executed on the host, needs to be prepended with "sudo <sudo-options>"
manually.

'sudo' is a shell script that is meant to be used before /bin/sudo to
make it possible to run commands on the host from inside of a container
with "sudo" prefixed. It checks first if the command, that is to be run
with sudo, is to be run on the host by checking if it is a symlink to
/usr/libexec/toolbox/host-runner. If it is not, it is executed normally
with /bin/sudo.

Both of these files should be put in /usr/libexec/toolbox which will
make them available in /run/host/usr/libexec/toolbox in toolbox
containers. Thanks to that there's no need to recreate already existing
containers because those should have the host /usr already available.