Deprecate the '--monitor-host' option of 'init-container'
The '--monitor-host' option was added to the 'init-container' command
in commit 8b84b5e4604921fa to accommodate Podman versions older than
1.2.0 that didn't have the '--dns none' and '--no-hosts' options for
'podman create'. These options are necessary to keep the toolbox
container's /etc/resolv.conf and /etc/hosts files synchronized with
those of the host.

Note that Podman 1.2.0 was already available a few months before
commit 8b84b5e4604921fa introduced the '--monitor-host' option. The
chances of someone using an older Podman back then was already on the
decline, and it's very unlikely that a container created such a Podman
has survived till this date.

Since commit b6b484fa792b442a, the minimum required Podman version is
1.4.0, and the '--dns none' and '--no-hosts' options are a hard
requirement. Therefore, these days, there's no need to separately use
the '--monitor-host' option of 'init-container' for newly created
containers to indicate that the Podman version wasn't older than 1.2.0.

Given all this, it's time to stop using the '--monitor-host' option of
'init-container', and assume that certain configuration files inside
the toolbox container should always be kept synchronized with their
counterparts on the host. The option is still accepted to retain
compatibility with existing toolbox containers.

https://github.com/containers/toolbox/pull/617